+++ changelog.md
@@ -1,7 +1,14 @@
 changelog:
+    11/15/22 - Crawler 0.6
+        - Fixed some bugs regarding map generation
+        - simplified the info command regarding avaliable doors to traverse
+        - added the ability to run from an enemy
+        - fixed a bug regarding a sleeping enemy and not being able to leave it alone.
+        - made things look prettier!
     11/14/22 - crawler 0.5
         - reworked the map generation so now it is truly random and not a sqare
         - added the feature where enemies may be asleep when you enter a room and wake when/if you leave the room
         - fixed some bugs regarding the stat command.
         - fixed bugs regarding the exit commands 
         - general bug fixes 
+        - made things look pretty-ish